Major bug is now fixed!
There has been a problem with getting some data from games like dota2. This was caused by the context id not being set to an int value. That is now fixed! Use string or an int, doesn't matter!

I also have set a callback function when we couldn't get the items from the inventory because the inventory was empty.

:+1: